Use a byte instead of an int as the amount of seek actions in the pending action queue

+++ osu.Framework/Audio/Track/TrackBass.cs
@@ -41,7 +41,7 @@ namespace osu.Framework.Audio.Track
 
         public override bool IsLoaded => isLoaded;
 
-        private volatile int seekCommands;
+        private volatile byte seekCommands;
 
         private readonly object seekLock = new Object();
